+++ tests/test_bit_container_size.rs
@@ -18,19 +18,19 @@ struct Test {
 
 #[rstest(input,
     #[should_panic(
-        expected = "InvalidParam(\"bit size of input is larger than bit requested size: 5 exceeds 4\")"
+        expected = "bit size of input is larger than bit requested size"
     )]
     case::field_u8_be( Test { field_u8_be: 0b11111, ..Test::default()}),
     #[should_panic(
-        expected = "InvalidParam(\"bit size of input is larger than bit requested size: 5 exceeds 4\")"
+        expected = "bit size of input is larger than bit requested size"
     )]
     case::field_be( Test { field_be: 0b11111, ..Test::default()}),
     #[should_panic(
-        expected = "InvalidParam(\"bit size of input is larger than requested size: 13 exceeds 12\")"
+        expected = "bit size of input is larger than requested size"
     )]
     case::field_le( Test { field_le: 0b1111111111111, ..Test::default()}),
     #[should_panic(
-        expected = "InvalidParam(\"bit size of input is larger than bit requested size: 13 exceeds 9\")"
+        expected = "bit size of input is larger than bit requested size"
     )]
     case::field_u32_be( Test { field_u32_be: 0b1111111111111, ..Test::default()}),
 )]