+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - The `alloc` feature, allowing use in environments lacking a heap [#582](https://github.com/sharksforarms/deku/pull/582)
+- The `descriptive-errors` feature, replacing `no-assertion-string` [#582](https://github.com/sharksforarms/deku/pull/582)
 
 ## [0.19.1] - 2025-05-22